Make the game partially work in the browser
I had to rewrite logic responsible for playing the game. It worked fine
in the console version and with player input obtained via the prompt()
method, but failed miserably in the browser, the reason being that
I could not pause code execution to wait for the user to click any
board square. The problem is solved now, luckily.
At this stage, the game works perfectly in a human vs human scenario
and kind of works for the remaining scenarios, but each computer move
must be made by clicking an empty square.